Fix int32 overflow in rotating-object back-face culling
The visibility dot product objX*nx + objY*ny + objZ*nz was computed in int.
With camera-relative coordinates around 1e9 (e.g. the rock pipeline at
yPlayer down to -52 tiles) and rotated normals around 1.8e8, the products
reach ~1e17 — far past int32 — so the wrapped sign was effectively random and
rotating objects (falling rocks, pyramids) culled faces unpredictably as they
tumbled.

The original avoids this by pre-scaling object coordinates to stay inside
32-bit products before the dot product (Lander.arm:5024-5081). The fix keeps
the intent — decide visibility by the dot's sign — but computes the exact
64-bit dot, matching PlayerController.cs which already does the same for the
ship's culling.

Adds a regression test with a synthetic rotating blueprint whose normal and
position are chosen so the product is provably negative but wraps to exactly 0
in int32 (0x3B9ACA00 * -0x50000000 = -0x12A05F2000000000, low 32 bits zero):
the face must be drawn, and the test fails against the int version.

+++ Relander.Core/Engine/ObjectRenderer.cs
@@ -24,8 +24,12 @@ public static class ObjectRenderer
                 ny = DotMatrix(face.Normal.X, face.Normal.Y, face.Normal.Z, 1, state);
                 nz = DotMatrix(face.Normal.X, face.Normal.Y, face.Normal.Z, 2, state);
 
-                // Back-face culling for rotating objects
-                int dot = objX * nx + objY * ny + objZ * nz;
+                // Back-face culling for rotating objects. The products exceed
+                // int32 (objX and normals reach ~1e9), so compute the exact sign
+                // in 64-bit — the original pre-scales coordinates to keep the
+                // products inside 32 bits (Lander.arm:5024-5081). The previous
+                // int arithmetic wrapped and culled faces at random.
+                long dot = (long)objX * nx + (long)objY * ny + (long)objZ * nz;
                 if (dot >= 0) continue;
             }
             else
